pyproject.toml
[project]
name = "cli-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.2.25",
    "google-generativeai>=0.8.6",
    "gradio>=6.9.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
